[EMACS] Simplify Info menu node sorting
Specify an ivy sorting function for the relevant collection function
which simply reverses the order of the parsed menu nodes. Compared
to the previous custom implementation, this results in a slightly
less accurate ordering relative to the final Info menu listing, but
the discrepancies are not critical and probably due to some Info
feature/bug.