[EMACS] Remove project-external-roots from Ibuffer
By default project-vc-external-roots-function includes
package-user-dir and the entire Elisp load-path under
emacs-lisp-mode.

That makes project-external-roots too context-sensitive and not
project-specific enough for Ibuffer filter groups, so revert
commit 9463304fb49b7b8e0bb003a4d51776f31f4ff0d9.

The original motivation was projects which have out-of-tree build
directories with useful artefacts.  The build directory was included
as an external rather than secondary root, since the latter was
deprecated in project.el.  I no longer use a workflow like that, so
I'll cross that bridge when I reach it again.